The exp basically worked, the quiz script is limmiting me.
Basically my idea was that we can spawn the ATMs in each level when the player is on that level, in order to not have more than 3 scripts for the Questions. It works basically but I can fully test the idea because whenever we press Play Again, it reloads the scene or something, and everything basically resets. We need to find where the developer of the Quiz Game fiddles with the play again button